Refactored the plate svg so it can be shared
