+++ src/components/ont/OntWell.vue
@@ -1,9 +1,7 @@
 <template>
-
   <ellipse v-bind:class="status" :cx="cx" :cy="cy" :rx="rx" :ry="ry" >
     <title v-if="hasMaterial" v-text="tooltip"></title>
   </ellipse>
-
 </template>
 
 <script>
@@ -26,7 +24,6 @@ export default {
 </script>
 
 <style scoped lang="scss">
-
   ellipse {
     transform: matrix(0.91863074,0,0,0.92029059,955.85411,1007.3112);
     stroke: #000000;
